Removed assumption that \\ will be present
The function zend_add_ns_func_name_literal is called if the parser finds a function that is not in the global or current namespace. It assumes such a function will have a \\ in it, which is no longer true with the use function patch. The code change above removes that assumption and makes the test work:
PASS use and use function with the same alias [Zend/tests/use_function/conflicting_use_alias.phpt]