Fix naming
The global section is named "General", but the correct section name is "global". It is "global" in #1313 and also "global" in https://github.com/php/php-src/blob/master/sapi/fpm/php-fpm.conf.in. The commit fixes the naming. However, the proposed change might break any tooling that relies on finding "General" in the output though.